+++ web/src/routes/admin/settings/+page.svelte
@@ -1,0 +1,7 @@
+
+
+
+
+<div>
+  <h1>Setting page</h1>
+</div>
